config: FileProvider.GoogleClientConfig should be able to read both client configs
google.ConfigFromJSON reads a configuration file which comes straight
from Google. It then turns it into an *oauth2.Config. When serializing
the *oauth2.Config, the resulting JSON is different than that which
comes from the Google Developer Console. It's most efficient to simply
deserialize the *oauth2.Config. To be backwards-compatible, we should
still try google.ConfigFromJSON before deserializing as an
*oauth2.Config.