zoom: extract the zoom URL and its password more robustly
The regexp approach has worked well! It was a bit rigid, however. When I went to ask it to include the query string so we could include the `pwd` entry, it felt like time to use a more robust library for extracting URLs, mvdan/xurls.

xurls can extract any URL from arbitrary user input, and relies on the URL spec and a fairly well-structure grammar for doing so.

This PR uses xurls to extract all URLs, then looks for the zoom.us URLs within that set, and pulls out the ID and password into a struct to then convert into a zoommtg:// URL.