Filter meeting list by closest to present time and beyond
When we query the Calendar API for a list of meetings, we do so by end
time. What we then want to do is choose the meeting with the closest
start time to now, and use that as the new beginning of the meeting
list, i.e. removing from the list all meetings that started prior to it.

This means that a meeting list like:

1. A meeting starting 30 mins ago
2. A meeting starting 5 mins ago
3. A meeting starting 30 mins from now

will result in a new list like:

1. A meeting starting 5 mins ago
2. A meeting starting 30 mins from now

The zoom launch will then open the meeting starting 5 mins ago, as
its closeness to the present time indicates its relevance.